isolate the whole-graph scan backstop into its own test binary
whole_graph_scan_limit() reads FLUREE_MAX_GRAPH_SCAN_FLAKES from the process
environment on every scan and has no programmatic override, so the backstop test
must set that variable to exercise the cap. Under bare `cargo test` a binary
runs its tests as threads in one process, so the cap applied between the set_var
and the remove_var leaked into whichever graph-sync tests happened to be running
alongside. delta_sync_commits_only_the_delta, identical_resync_is_a_noop,
sync_does_not_touch_other_graphs and policy_gated_sync_targets_the_named_graph
all failed with WholeGraphScanTooLarge { limit: 2 } depending on scheduling.

Move that one test into tests/it_sync_graph_scan_backstop.rs with its own
[[test]] target; the remaining graph-sync tests stay in grp_ledger. Keeping all
the env mutation inside a single test fn is not enough on its own. That stops
the mutators racing each other, not the mutation reaching siblings in the same
binary.

nextest gives every test its own process, which is why CI has always been green
and why this only reproduces with `cargo test -p fluree-db-api --test grp_ledger`.
