+++ fluree-db-api/tests/it_sync_graph.rs
@@ -540,71 +540,6 @@ async fn identical_resync_is_a_noop_against_indexed_data() {
         .await;
 }
 
-/// The whole-graph memory backstop: staging materializes the target
-/// graph's current flakes, so a graph past the cap must fail loud (an
-/// OOM kill otherwise, with no guard in between — `NoveltyWouldExceed`
-/// only ever sees the netted delta). The limit is read per scan, so the
-/// env changes below take effect immediately; nextest's process-per-test
-/// isolation keeps them from leaking into other tests.
-#[tokio::test]
-async fn whole_graph_scan_backstop_fails_loud_before_materializing() {
-    let fluree = FlureeBuilder::memory().build_memory();
-    let ledger_id = "it/sync-graph/backstop:main";
-    seed(&fluree, ledger_id).await;
-
-    std::env::set_var("FLUREE_MAX_GRAPH_SCAN_FLAKES", "2");
-    // First sync scans an EMPTY graph (0 <= 2): the cap bounds the scan,
-    // not the payload, so populating past the cap succeeds.
-    let report = fluree
-        .sync_named_graph(ledger_id, ONT_IRI, &payload_v1(), SyncGraphOpts::default())
-        .await
-        .expect("first sync scans an empty graph");
-    assert_eq!(report.asserted, 3);
-
-    // Now the graph holds 3 > 2: every whole-graph verb refuses.
-    let err = fluree
-        .sync_named_graph(ledger_id, ONT_IRI, &payload_v1(), SyncGraphOpts::default())
-        .await
-        .expect_err("resync over the cap must fail loud");
-    assert!(
-        err.to_string().contains("FLUREE_MAX_GRAPH_SCAN_FLAKES"),
-        "error names the knob: {err}"
-    );
-
-    // Dry run takes the same scan and must fail the same way.
-    let dry = fluree
-        .sync_named_graph(
-            ledger_id,
-            ONT_IRI,
-            &payload_v1(),
-            SyncGraphOpts {
-                dry_run: true,
-                ..Default::default()
-            },
-        )
-        .await
-        .expect_err("dry run must fail the way the real run would");
-    assert!(dry.to_string().contains("FLUREE_MAX_GRAPH_SCAN_FLAKES"));
-
-    let ledger = fluree.ledger(ledger_id).await.unwrap();
-    let clear = fluree
-        .stage_owned(ledger)
-        .txn(fluree_db_transact::Txn::clear_graph(ONT_IRI))
-        .execute()
-        .await
-        .expect_err("CLEAR shares the scan and the backstop");
-    assert!(clear.to_string().contains("FLUREE_MAX_GRAPH_SCAN_FLAKES"));
-
-    // 0 disables; the identical resync is a no-op again.
-    std::env::set_var("FLUREE_MAX_GRAPH_SCAN_FLAKES", "0");
-    let report = fluree
-        .sync_named_graph(ledger_id, ONT_IRI, &payload_v1(), SyncGraphOpts::default())
-        .await
-        .expect("disabled cap syncs normally");
-    assert!(!report.committed);
-    std::env::remove_var("FLUREE_MAX_GRAPH_SCAN_FLAKES");
-}
-
 /// CLEAR / COPY / MOVE on an index-resident named graph — the same scan sync
 /// rides, and broken the same two ways on `main` before this branch: the
 /// scan issued `RangeTest::Ge`, which the V3 provider rejects, and once it
